pyproject.toml
[project]
name = "data-exporter"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "requests>=2.31",
    "boto3>=1.34",
    "google-api-python-client>=2.100",
    "google-auth>=2.23",
    "google-auth-httplib2>=0.2",
    "PyJWT[crypto]>=2.8",
    "google-cloud-bigquery>=3.20",
    "google-cloud-bigquery-storage>=2.24",
    "pyarrow>=15.0",
    "orjson>=3.11.7",
    "duckdb>=1.5.1",
    "Faker>=33.0",
    "pyahocorasick>=2.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-mock>=3.12",
    "moto[s3]>=5.0",
    "responses>=0.25",
]
fast = [
    "boto3[crt]",
]
ner = [
    "presidio-analyzer>=2.2",
    "presidio-anonymizer>=2.2",
    "spacy>=3.7",
]

[tool.hatch.build.targets.wheel]
packages = ["lib", "exporters", "scripts"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
